Basic caching, log is now RotatingFileHandler.
Turned some basic db-caching on for pages in projects, and misc.
Still needs further testing on a live site, cache_dummy is used
for local dev.

Switched log file handler to RotatingFileHandler.
This breaks Python 2 compatibility! No looking back now.
As of right now the log file is rotated when it reaches 2MB,
and no backups are made.