Added no-access to test site content for non-admin
You must login before viewing the test site.
Somehow people got a hold of an old test-url,
and tried to access it on the live site.
It did not harm anything, a 404 was returned.
I don't want people grabbing info from the
test site to possibly use on the live site.