Switching to wrun.sh, instead of wrun.py.
The python script has trouble when the child scripts
need input from the user. The work-around was ugly and
required updating 'wrun.py' every time a new tool was added.
(when that tool requires user input)

The bash wrun doesn't have this problem, and does a fine job
of running scripts. (isn't that was bash script are for?)

Time will tell if this was a mistake, but I don't foresee any
major problems.