New custom Github badge added.
The old one from http://githubbadge.appspot.com/ fails
on organizations, and flat-out errors when the user has
no repos (leaving an iframe with ERROR written on it).

This badge is designed to fail gracefully. If the info
can't be retrieved the div will be hidden. I drew a lot
of inspiration from the old one, but kept the colors
from my main sass file.