Overhaul updateobject, apps now need an update.py.
Apps that have an `update.py` in their package can
now be updated from the command line.
The `update.py` script in (`./scripts/`) will look
for any app with that module, and grab some needed info
for working with the app's models.

TODO: Apps may have more than one model in the future.
      It would be nice if update.py could work like that.