Use lxml to parse pre tags when highlighting.
The old "state machine" method was a big ugly hack.
With this method, lxml is doing the parsing and I'm
just replacing the pre tag's text with some html
(highlighted text). The highlight code is still a
little ugly because of special-case classes like
'none', 'codewrap', and 'codewrap-template', which
all do something different and negate the normal
<pre class='language'> flow.